pyproject.toml
[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "feature-platform"
version = "0.1.0"
description = "A feature platform for managing and serving machine learning features"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "databricks-sql-connector>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=0.910",
    "flake8>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=feature_platform --cov-report=term-missing"

[project.urls]
Homepage = "https://github.com/yourusername/feature-platform"
Documentation = "https://github.com/yourusername/feature-platform#readme"
Issues = "https://github.com/yourusername/feature-platform/issues"

[tool.coverage.run]
source = ["domain"]
omit = ["**/__init__.py"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 90
